Full refactor Frontend
