+++ frontend/src/components/DoughForm/index.ts
@@ -1,0 +1,1 @@
+export { DoughForm } from './DoughForm'
