+++ frontend/src/components/DoughResultBreakdown/index.ts
@@ -1,0 +1,1 @@
+export { DoughResultBreakdown } from './DoughResultBreakdown'
