+++ frontend/src/components/NumberField/index.ts
@@ -1,0 +1,1 @@
+export { NumberField } from './NumberField'
