+++ frontend/src/components/RecipeComparisonView/index.ts
@@ -1,0 +1,1 @@
+export { RecipeComparisonView } from './RecipeComparisonView'
