+++ frontend/src/components/RecipeDetailModal/RecipeDetailModal.scss
@@ -1,0 +1,191 @@
+@use '../../styles/tokens' as *;
+
+.recipe-modal-backdrop {
+  position: fixed;
+  inset: 0;
+  z-index: 20;
+  display: grid;
+  place-items: center;
+  padding: 18px;
+  background: $color-surface-overlay;
+}
+
+.recipe-modal {
+  width: min(720px, 100%);
+  max-height: min(720px, calc(100vh - 36px));
+  overflow: auto;
+  padding: 18px;
+  border-radius: $radius-md;
+  background: $color-surface-card-strong;
+  box-shadow: $shadow-modal;
+}
+
+.recipe-modal-header {
+  display: flex;
+  align-items: start;
+  justify-content: space-between;
+  gap: 16px;
+  margin-bottom: 16px;
+}
+
+.recipe-modal-header h2 {
+  color: $color-text-strong;
+  font-size: 32px;
+  line-height: 1.1;
+}
+
+.recipe-modal-header button {
+  min-width: 82px;
+  min-height: 40px;
+  border: 0;
+  border-radius: $radius-sm;
+  background: $color-surface-muted;
+  color: $color-brand;
+  font: inherit;
+  font-weight: 800;
+  cursor: pointer;
+}
+
+.recipe-modal-grid {
+  display: grid;
+  grid-template-columns: 1fr 1fr;
+  gap: 12px;
+}
+
+.recipe-modal-form {
+  display: grid;
+  gap: 18px;
+  margin-bottom: 16px;
+  padding: 18px;
+  border: 1px solid $color-border-default;
+  border-radius: $radius-md;
+  background: $color-surface-card;
+}
+
+.recipe-modal-toolbar {
+  display: grid;
+  gap: 12px;
+  margin-bottom: 12px;
+}
+
+.recipe-modal-toolbar > button,
+.recipe-modal-action-row button {
+  min-height: 42px;
+  border: 0;
+  border-radius: $radius-sm;
+  background: $color-brand;
+  color: #fff;
+  font: inherit;
+  font-weight: 800;
+  cursor: pointer;
+}
+
+.recipe-modal-toolbar > button:nth-child(2),
+.recipe-modal-action-row button:last-child {
+  background: $color-surface-muted;
+  color: $color-brand;
+}
+
+.recipe-modal-toolbar > button:last-child {
+  background: $color-surface-danger;
+  color: $color-brand-strong;
+}
+
+.recipe-modal-name-field {
+  display: grid;
+  gap: 7px;
+  color: $color-text-soft;
+  font-size: 13px;
+  font-weight: 700;
+}
+
+.recipe-modal-name-field input {
+  min-height: 44px;
+  min-width: 0;
+  padding: 0 10px;
+  border: 1px solid $color-border-strong;
+  border-radius: $radius-sm;
+  background: $color-surface-card;
+  color: $color-text;
+  font: inherit;
+  font-size: 16px;
+  font-weight: 800;
+  outline: 0;
+}
+
+.recipe-modal-action-row {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 8px;
+}
+
+.recipe-modal-hint {
+  margin: 0 0 16px;
+  color: $color-text-soft;
+  font-size: 14px;
+  line-height: 1.5;
+}
+
+.recipe-summary-block {
+  padding: 14px;
+  border: 1px solid $color-border-default;
+  border-radius: $radius-md;
+  background: $color-surface-card;
+}
+
+.recipe-summary-block h3 {
+  margin-bottom: 10px;
+  color: $color-text-accent;
+  font-size: 18px;
+}
+
+.recipe-facts {
+  display: grid;
+  gap: 8px;
+  margin: 0;
+}
+
+.recipe-facts div {
+  display: flex;
+  justify-content: space-between;
+  gap: 12px;
+  padding-bottom: 8px;
+  border-bottom: 1px solid $color-border-default;
+}
+
+.recipe-facts div:last-child {
+  padding-bottom: 0;
+  border-bottom: 0;
+}
+
+.recipe-facts dt {
+  color: $color-text-muted;
+  font-weight: 700;
+}
+
+.recipe-facts dd {
+  margin: 0;
+  color: $color-text-strong;
+  font-weight: 800;
+  text-align: right;
+}
+
+@media (max-width: $breakpoint-tablet) {
+  .recipe-modal {
+    padding: 14px;
+  }
+
+  .recipe-modal-header {
+    align-items: stretch;
+    flex-direction: column;
+  }
+
+  .recipe-modal-grid {
+    grid-template-columns: 1fr;
+  }
+
+  .recipe-modal-header button,
+  .ingredient-list-modal {
+    width: 100%;
+  }
+}
