+++ frontend/src/components/RecipeListItem/RecipeListItem.scss
@@ -1,0 +1,105 @@
+@use '../../styles/tokens' as *;
+
+.recipe-item {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 12px;
+  min-height: 68px;
+  padding: 12px 14px;
+  border: 1px solid $color-border-default;
+  border-radius: $radius-md;
+  background: $color-surface-card;
+  cursor: pointer;
+  transition: $transition-fast;
+}
+
+.recipe-item:hover {
+  border-color: $color-border-hover;
+  box-shadow: $shadow-card-hover;
+  transform: translateY(-1px);
+}
+
+.recipe-item.active {
+  border-color: $color-brand;
+  box-shadow: $shadow-focus;
+}
+
+.recipe-item.disabled {
+  cursor: wait;
+}
+
+.recipe-item:focus-visible {
+  outline: 2px solid $color-brand;
+  outline-offset: 2px;
+}
+
+.recipe-item-body {
+  min-width: 0;
+}
+
+.recipe-item h3 {
+  color: $color-text-strong;
+  font-size: 18px;
+}
+
+.recipe-item p {
+  color: $color-text-muted;
+  font-size: 14px;
+}
+
+.recipe-actions,
+.recipe-ingredients {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 8px;
+}
+
+.recipe-actions {
+  justify-content: flex-end;
+}
+
+.recipe-actions button {
+  min-width: 76px;
+  min-height: 42px;
+  border: 0;
+  border-radius: $radius-sm;
+  background: $color-surface-muted;
+  color: $color-brand;
+  font: inherit;
+  font-weight: 800;
+  cursor: pointer;
+}
+
+.recipe-actions button:disabled {
+  cursor: wait;
+  opacity: 0.7;
+}
+
+.recipe-ingredients {
+  margin-top: 8px;
+}
+
+.recipe-ingredients span {
+  padding: 6px 8px;
+  border-radius: $radius-pill;
+  background: $color-surface-accent;
+  color: $color-brand-soft;
+  font-size: 13px;
+  font-weight: 800;
+}
+
+@media (max-width: $breakpoint-tablet) {
+  .recipe-item {
+    align-items: stretch;
+    flex-direction: column;
+  }
+
+  .recipe-actions {
+    justify-content: stretch;
+  }
+
+  .recipe-actions button {
+    width: 100%;
+  }
+}
