+++ frontend/src/components/RecipeListItem/index.ts
@@ -1,0 +1,1 @@
+export { RecipeListItem } from './RecipeListItem'
