+++ frontend/src/components/RecipeManager/index.ts
@@ -1,0 +1,1 @@
+export { RecipeManager } from './RecipeManager'
