+++ frontend/src/components/ResultsPanel/index.ts
@@ -1,0 +1,1 @@
+export { ResultsPanel } from './ResultsPanel'
