+++ frontend/src/styles/global.scss
@@ -1,0 +1,65 @@
+@use './tokens' as *;
+
+:root {
+  font-family:
+    Inter, ui-sans-serif, system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', sans-serif;
+  color: $color-text;
+  background: $color-surface-page;
+  font-synthesis: none;
+  text-rendering: optimizeLegibility;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+}
+
+* {
+  box-sizing: border-box;
+}
+
+body {
+  margin: 0;
+  min-width: 320px;
+}
+
+button,
+input,
+select {
+  letter-spacing: 0;
+}
+
+h1,
+h2,
+h3,
+h4,
+p {
+  margin: 0;
+}
+
+.eyebrow {
+  margin: 0 0 6px;
+  color: $color-brand;
+  font-size: 13px;
+  font-weight: 800;
+  letter-spacing: 0;
+  text-transform: uppercase;
+}
+
+.section-title {
+  color: $color-text-accent;
+  font-size: 13px;
+  font-weight: 800;
+  text-transform: uppercase;
+}
+
+.error-message {
+  padding: 10px 12px;
+  border-radius: $radius-sm;
+  background: $color-surface-danger;
+  color: $color-brand-strong;
+  font-size: 14px;
+}
+
+.empty-recipes {
+  padding: 12px 0;
+  color: $color-text-muted;
+  font-size: 14px;
+}
