+++ frontend/src/utils/format.ts
@@ -1,0 +1,3 @@
+export function formatGram(value: number) {
+  return `${value.toFixed(1)}g`
+}
